test: extract tree model from hypothesis importing file
